+++ packages/cli/demo/typescript/es.ts
@@ -1,9 +1,0 @@
-export default {
-  title: "Hola Mundo",
-  description: "Esta es una descripción",
-  field1: "Hola",
-  field2: "Mundo",
-  field3: {
-    field4: "Hola Mundo"
-  }
-} as const;