+++ packages/react/src/core/const.ts
@@ -1,9 +1,4 @@
 /**
- * The default locale.
- */
-export const DEFAULT_LOCALE = "en";
-
-/**
  * The name of the cookie that stores the current locale.
  */
 export const LOCALE_COOKIE_NAME = "lingo-locale";