+++ packages/react/src/core/index.ts
@@ -1,3 +1,4 @@
 export * from "./component";
 export * from "./const";
 export * from "./attribute-component";
+export * from "./get-dictionary";