+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @lingo.dev/\_sdk
 
+## 0.12.0
+
+### Minor Changes
+
+- [#1066](https://github.com/lingodotdev/lingo.dev/pull/1066) [`6af91a0`](https://github.com/lingodotdev/lingo.dev/commit/6af91a083d16f85051fb49a4034789abe784017e) Thanks [@maxprilutskiy](https://github.com/maxprilutskiy)! - add hints support
+
+### Patch Changes
+
+- Updated dependencies [[`6af91a0`](https://github.com/lingodotdev/lingo.dev/commit/6af91a083d16f85051fb49a4034789abe784017e)]:
+  - @lingo.dev/_spec@0.40.0
+
+## 0.11.0
+
+### Minor Changes
+
+- [#1049](https://github.com/lingodotdev/lingo.dev/pull/1049) [`85dfc10`](https://github.com/lingodotdev/lingo.dev/commit/85dfc10961b116e31b2bb478f42013756ca49974) Thanks [@VAIBHAVSING](https://github.com/VAIBHAVSING)! - Added new methods to the SDK:
+
+  1. `localizeStringArray`: Localizes an array of strings while maintaining their order.
+
+  Also added comprehensive tests for these methods using Vitest.
+
 ## 0.10.2
 
 ### Patch Changes